# keychain+lnwallet: when fetching priv keys or signing try to use cache 